+++ test/transpile/transformers/fixtures.ts
@@ -1,3 +1,4 @@
+import * as assert from "assert";
 import * as ts from "typescript";
 import * as tstl from "../../../src";
 import { visitAndReplace } from "./utils";
@@ -7,7 +8,7 @@ export const program = (program: ts.Program, options: { value: any }): ts.Transf
 
 export const config = ({ value }: { value: any }): ts.TransformerFactory<ts.SourceFile> => context => file =>
     visitAndReplace(context, file, node => {
-        if (!ts.isReturnStatement(node) || node.expression) return;
+        if (!ts.isReturnStatement(node)) return;
         return ts.updateReturn(node, ts.createLiteral(value));
     });
 
@@ -24,19 +25,16 @@ export const checker = (
 
 export const raw: ts.TransformerFactory<ts.SourceFile> = context => file =>
     visitAndReplace(context, file, node => {
-        if (!ts.isReturnStatement(node) || node.expression) return;
+        if (!ts.isReturnStatement(node)) return;
         return ts.updateReturn(node, ts.createLiteral(true));
     });
 
-export const compilerOptions = ({
-    luaTarget,
-}: tstl.CompilerOptions): ts.TransformerFactory<ts.SourceFile> => context => file => {
-    if (luaTarget !== tstl.LuaTarget.LuaJIT) {
-        throw new Error("Transformer supports only LuaJIT target");
-    }
-
+export const compilerOptions = (
+    options: tstl.CompilerOptions
+): ts.TransformerFactory<ts.SourceFile> => context => file => {
+    assert(options.plugins?.length === 1);
     return visitAndReplace(context, file, node => {
-        if (!ts.isReturnStatement(node) || node.expression) return;
+        if (!ts.isReturnStatement(node)) return;
         return ts.updateReturn(node, ts.createLiteral(true));
     });
 };