+++ CHANGELOG.md
@@ -81,6 +81,16 @@
   end
   ```
 
+- Added `tstl.luaPlugins` option, allowing to specify plugins in a `tsconfig.json` file:
+
+  ```json
+  {
+    "tstl": {
+      "luaPlugins": [{ "name": "./plugin.ts" }]
+    }
+  }
+  ```
+
 ## 0.31.0
 
 - **Breaking:** The old annotation syntax (`/* !varArg */`) **no longer works**, the only currently supported syntax is: